+++ app/components/base/table/index.ts
@@ -1,6 +1,6 @@
 export * from "./table.component";
-export * from "./table-head.component";
-export * from "./table-row.component";
-export * from "./table-cell.component";
-export * from "./table-column.component";
+export * from "./table-head";
+export * from "./table-row";
+export * from "./table-cell";
+export * from "./table-column";
 export * from "./table.module";
