+++ app/components/base/table/table-cell/index.ts
@@ -1,0 +1,1 @@
+export * from "./table-cell.component";
