+++ app/components/base/table/table-column/index.ts
@@ -1,0 +1,1 @@
+export * from "./table-column.component";
