+++ app/components/base/table/table-head/index.ts
@@ -1,0 +1,1 @@
+export * from "./table-head.component";
