+++ app/components/base/table/table-row/index.ts
@@ -1,0 +1,1 @@
+export * from "./table-row.component";
