+++ app/components/base/table/table.module.ts
@@ -1,14 +1,14 @@
-import { ModuleWithProviders, NgModule } from "@angular/core";
+import { NgModule } from "@angular/core";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { BrowserModule } from "@angular/platform-browser";
 import { RouterModule } from "@angular/router";
 import { MaterialModule } from "app/core";
 
 import { FocusSectionModule } from "../focus-section";
-import { TableCellComponent } from "./table-cell.component";
-import { TableColumnComponent } from "./table-column.component";
-import { TableHeadComponent } from "./table-head.component";
-import { TableRowComponent } from "./table-row.component";
+import { TableCellComponent } from "./table-cell";
+import { TableColumnComponent } from "./table-column";
+import { TableHeadComponent } from "./table-head";
+import { TableRowComponent } from "./table-row";
 import { TableComponent } from "./table.component";
 
 const components = [
@@ -33,10 +33,4 @@ const components = [
     providers: [],
 })
 export class TableModule {
-    public static forRoot(): ModuleWithProviders {
-        return {
-            ngModule: TableModule,
-            providers: [],
-        };
-    }
 }
